requirements.txt
psycopg2-binary==2.9.1

tksheet~=5.2.8